Minor fixes and changes
Changed the url for the existing api from */api/* to */api/user/* to better reflect its use.
Fixed a bug where the User Context was not updating when logging in, out or registering.
Made the username display use the User Context.
Changed the user context from layoutClient.tsx to layout.tsx.
Fixed a bug where the user was being logged in twice, causing errors